Add files via upload
The primary goal of this project is to leverage the capabilities of Apache Spark, specifically PySpark, to perform comprehensive analysis on merged datasets of temperature, pressure, and humidity. By utilizing the MapReduce programming model within a distributed computing environment, the project aims to efficiently process and analyze large volumes of meteorological data to uncover patterns and insights that are crucial for climatic research and weather prediction.